fix: tighten desktop startup package checks
