+++ packages/desktop/src/main/index.ts
@@ -31,6 +31,7 @@ import {
   type DesktopAuthCallbackState,
   type DesktopAuthCoordinator,
 } from './desktopSupabaseAuth.js';
+import { reportFatalStartupError } from './fatalStartupError.js';
 import { installDesktopMainViewIpc } from './mainViewIpc.js';
 import { initializeElectronPlatform } from './platform/index.js';
 import {
@@ -337,9 +338,7 @@ if (protocolLifecycle.shouldContinue) {
       });
     })
     .catch((error: unknown) => {
-      const message = error instanceof Error ? error.message : String(error);
-      console.error(`Failed to start TeXRA desktop: ${message}`);
-      app.quit();
+      reportFatalStartupError(error);
     });
 }
